Set maximum resource limits when using Fusion Snapshots
Some customers have reported issues with asking for too much memory or CPUs when using Fusion Snapshots due to the limited resources of the instance types available.

This commit adds a `process.resourceLimits` setting to the Nextflow config to limit the maximum resource usage of the task to fit within the restricted instance types.